+++ packages/react-form/src/hooks/list/index.ts
@@ -1,2 +1,4 @@
 export {useList} from './list';
 export {useDynamicList} from './dynamiclist';
+export {useDynamicListDirty} from './dynamiclistdirty';
+export {useDynamicListReset} from './dynamiclistreset';