+++ packages/react-form/src/hooks/test/utilities/index.ts
@@ -1,0 +1,68 @@
+import faker from 'faker';
+
+import {SimpleProduct, TextField} from './components';
+
+export * from './components';
+
+export function isDirty(wrapper) {
+  try {
+    expect(wrapper).toContainReactComponent('button', {
+      type: 'reset',
+      disabled: false,
+    });
+    expect(wrapper).toContainReactComponent('button', {
+      type: 'submit',
+      disabled: false,
+    });
+  } catch {
+    return false;
+  }
+  return true;
+}
+
+export function changeTitle(wrapper, newTitle) {
+  wrapper.find(TextField, {label: 'title'})!.trigger('onChange', newTitle);
+}
+
+export function hitSubmit(wrapper) {
+  wrapper.find('button', {type: 'submit'})!.trigger('onClick', clickEvent());
+}
+
+export async function waitForSubmit(wrapper, successPromise) {
+  hitSubmit(wrapper);
+
+  await wrapper.act(async () => {
+    await successPromise;
+  });
+}
+
+export function hitReset(wrapper) {
+  wrapper.find('button', {type: 'reset'})!.trigger('onClick', clickEvent());
+}
+
+export function hitClean(wrapper) {
+  wrapper.find('button', {type: 'button'})!.trigger('onClick', clickEvent());
+}
+
+export function fakeProduct(): SimpleProduct {
+  return {
+    title: faker.commerce.product(),
+    description: faker.lorem.paragraph(),
+    defaultVariant: {
+      price: faker.commerce.price(),
+      optionName: 'material',
+      optionValue: faker.commerce.productMaterial(),
+    },
+    variants: Array.from({length: 2}).map(() => ({
+      id: faker.random.uuid(),
+      price: faker.commerce.price(),
+      optionName: faker.lorem.word(),
+      optionValue: faker.commerce.productMaterial(),
+    })),
+  };
+}
+
+export function clickEvent() {
+  // we don't actually use these at all so it is ok to just return an empty object
+  return {} as any;
+}